requirements.txt
aiogram
aiohttp
anitopy
Pillow
psutil
hachoir
tgcrypto
pyrogram==2.0.106
motor
https://github.com/pyrogram/pyrogram/archive/master.zip
dnspython
requests
speedtest-cli
